♻️[REFACTOR] : changeToDouble을 string extension에 구현
