+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -39,7 +39,5 @@ struct CalculatorItemQueue<T: CalculateItem> {
     }
 }
 
-extension Double: CalculateItem {
-}
